Models

(1+1)-dimensional Quantum Hamiltonians

MPSKitModels.transverse_field_ising — Function

transverse_field_ising([elt::Type{<:Number}], [symmetry::Type{<:Sector}],
                       [lattice::AbstractLattice]; J=1.0, g=1.0)

MPO for the hamiltonian of the Transverse-field Ising model, as defined by

\[H = -J\left(∑_{<i,j>} Z_i Z_j + g ∑_{<i>} X_i\right)\]

By default, the model is defined on an infinite chain with unit lattice spacing, without any symmetries and with ComplexF64 entries of the tensors.

MPSKitModels.kitaev_model — Function

kitaev_model([elt::Type{<:Number}], [lattice::AbstractLattice];
             t=1.0, mu=1.0, Delta=1.0)

MPO for the hamiltonian of the Kitaev model, as defined by

\[H = ∑_{<i,j>} \left(-\frac{t}{2}(c⁺ᵢcⱼ + c⁺ⱼcᵢ) + \frac{Δ}{2}(c⁺ᵢc⁺ⱼ + cⱼcᵢ) \right) - μ ∑_{<i>} c⁺ᵢcᵢ\]

By default, the model is defined on an infinite chain with unit lattice spacing and with ComplexF64 entries of the tensors.

MPSKitModels.heisenberg_XXX — Function

heisenberg_XXX([elt::Type{<:Number}], [symmetry::Type{<:Sector}],
                       [lattice::AbstractLattice]; J=1.0, spin=1)

MPO for the hamiltonian of the isotropic Heisenberg model, as defined by

\[H = J ∑_{<i,j>} S⃗ᵢ⋅S⃗ⱼ\]

By default, the model is defined on an infinite chain with unit lattice spacing, without any symmetries and with ComplexF64 entries of the tensors.

See also heisenberg_XXZ and heisenberg_XYZ.

MPSKitModels.heisenberg_XXZ — Function

heisenberg_XXZ([elt::Type{<:Number}], [symmetry::Type{<:Sector}],
               [lattice::AbstractLattice]; J=1.0, Delta=1.0, spin=1)

MPO for the hamiltonian of the XXZ Heisenberg model, as defined by

\[H = J(∑_{<i,j>} XᵢXⱼ + YᵢYⱼ + Δ ZᵢZⱼ)\]

By default, the model is defined on an infinite chain with unit lattice spacing, without any symmetries and with ComplexF64 entries of the tensors.

MPSKitModels.heisenberg_XYZ — Function

heisenberg_XYZ([elt::Type{<:Number}], [lattice::AbstractLattice];
    Jx=1.0, Jy=1.0, Jz=1.0, spin=1)

MPO for the hamiltonian of the xyz Heisenberg model, defined by

\[H = ∑_{<i,j>} (JˣXᵢXⱼ + JʸYᵢYⱼ + JᶻZᵢZⱼ)\]

By default, the model is defined on an infinite chain with unit lattice spacing and with ComplexF64 entries of the tensors.

MPSKitModels.bilinear_biquadratic_model — Function

bilinear_biquadratic_model([elt::Type{<:Number}], [symmetry::Type{<:Sector}],
                           [lattice::AbstractLattice]; spin=1, J=1.0, θ=0.0)

MPO for the hamiltonian of the bilinear biquadratic Heisenberg model, as defined by

\[H = J ∑_{<i,j>} (\cos(θ) S⃗ᵢ⋅S⃗ⱼ + \sin(θ) (⃗S⃗ᵢ⋅S⃗ⱼ)²)\]

By default, the model is defined on an infinite chain with unit lattice spacing, without any symmetries and with ComplexF64 entries of the tensors.

MPSKitModels.hubbard_model — Function

hubbard_model([elt::Type{<:Number}], [particle_symmetry::Type{<:Sector}],
              [spin_symmetry::Type{<:Sector}], [lattice::AbstractLattice];
              t, U, mu, n)

MPO for the hamiltonian of the Hubbard model, as defined by

\[H = -t∑_{<i,j>} (c⁺_{σ,i}c⁻_{σ,j} + c⁻_{σ,i}c⁺_{σ,j}) + U ∑_i n_{i,↑}n_{i,↓} - ∑_i μnᵢ\]

By default, the model is defined on an infinite chain with unit lattice spacing, without any symmetries and with ComplexF64 entries of the tensors. If the particle_symmetry is not Trivial, a fixed particle number density n can be imposed.

MPSKitModels.bose_hubbard_model — Function

bose_hubbard_model([elt::Type{<:Number}], [symmetry::Type{<:Sector}],
                   [lattice::AbstractLattice];
                   cutoff, t, U, mu, n)

MPO for the hamiltonian of the Bose-Hubbard model, as defined by

\[H = -t∑_{<i,j>} (a⁺_{i}a⁻_{j} + a⁻_{i}a⁺_{j}) - ∑_i μnᵢ + U / 2 ∑_i nᵢ(nᵢ - 1)\]

By default, the model is defined on an infinite chain with unit lattice spacing, without any symmetries and with ComplexF64 entries of the tensors. The Hilbert space is truncated such that at maximum of cutoff bosons can be at a single site. If the symmetry is not Trivial, a fixed particle number density n can be imposed.

MPSKitModels.quantum_chemistry_hamiltonian — Function

quantum_chemistry_hamiltonian(E0::Number, K::AbstractMatrix{<:Number}, V::AbstractArray{<:Number,4}, [elt::Type{<:Number}=ComplexF64])

MPO for the quantum chemistry Hamiltonian, with kinetic energy K and potential energy V. The Hamiltonian is given by

\[H = E0 + ∑ᵢⱼ ∑ₛ K[i,j] c^{s,+}_i c^{s,-}_j + ∑ᵢⱼₖₗ ∑ₛₜ V[i,j,k,l] c^{s,+}_i c^{t,+}_j c^{t,-}_k c^{s,-}_l\]

where s and t are spin indices, which can be ↑ or ↓. The full hamiltonian has U₁ × SU₂ × FermionParity symmetry.

Note

This should not be regarded as state-of-the-art quantum chemistry DMRG code. It is only meant to demonstrate the use of MPSKit for quantum chemistry. In particular:

No attempt was made to incorporate spacegroup symmetries

MPSKit does not contain many required algorithms in quantum chemistry (orbital ordering/optimization)

MPOHamiltonian is not well suited for quantum chemistry

(2+0)-dimensional Statistical Mechanics

MPSKitModels.classical_ising — Function

classical_ising([elt::Type{<:Number}=ComplexF64], [symmetry::Type{<:Sector}=Trivial];
                beta=log(1+sqrt(2))/2)

MPO for the classical Ising partition function, defined by

$Z(β) = ∑_s exp(-βH(s))$ with $H(s) = ∑_{<i,j>}σ_i σ_j$

MPSKitModels.sixvertex — Function

sixvertex([elt::Type{<:Number}=ComplexF64], [symmetry::Type{<:Sector}=Trivial];
          a=1.0, b=1.0, c=1.0)

MPO for the six vertex model.

MPSKitModels.hard_hexagon — Function

hard_hexagon([elt::Type{<:Number}=ComplexF64])

MPO for the hard hexagon model.

MPSKitModels.qstate_clock — Function

qstate_clock([elt::Type{<:Number}=ComplexF64], [symmetry::Type{<:Sector}=Trivial]; beta::Number=1.0, q::Integer=3)

MPO for the discrete clock model with $q$ states.