Models

MPSKitModels.transverse_field_ising — Function

transverse_field_ising(; J=1.0, hx=1.0, hz=0.0, spin=1//2)

MPO for the hamiltonian of the transverse field Ising model, defined by $H = -J(∑_{<i,j>} Z_i Z_j + ∑_{<i>} hx X_i + hz Z_i)$

Missing docstring.

Missing docstring for classical_ising. Check Documenter's build log for details.

Missing docstring.

Missing docstring for sixvertex. Check Documenter's build log for details.

MPSKitModels.xxx — Function

xxx(; J=1.0, spin=1)

MPO for the hamiltonian of the xxx Heisenberg model, defined by $H = J(∑_{<i,j>} X_i X_j + Y_i Y_j + Z_i Z_j)$

MPSKitModels.xxz — Function

xxz(; J=1.0, Δ=1.0, spin=1)

MPO for the hamiltonian of the xxz Heisenberg model, defined by $H = J(∑_{<i,j>} X_i X_j + Y_i Y_j + Δ Z_i Z_j)$

MPSKitModels.xyz — Function

xyz(; Jx=1.0, Jy=1.0, Jz=1.0, spin=1)

MPO for the hamiltonian of the xyz Heisenberg model, defined by $H = J(∑_{<i,j>} J_x X_{i}X_{j} + J_y Y_{i}Y_{j} + J_z Z_{i}Z_{j})$

MPSKitModels.su2su2_grossneveu — Function

su2su2_grossneveu(;g=0.,v=0.)

Returns the SU₂⊗SU₂ Gross-Neveu Hamiltonian in curved spacetime for uniform v(x)=v.

su2su2_grossneveu(vs::Vector{Float64};g=0.)

Returns the SU₂⊗SU₂ Gross-Neveu Hamiltonian in curved spacetime for non-uniform v(x). The array of numbers vs contains the value of v(x) sampled at discrete points. The length of vs determines the amount of sites

MPSKitModels.nonsym_qed_qlm_ham — Function

a quantum link model of qed

MPSKitModels.su2_xxx_ladder — Function

γ is the interchain coupling strength

MPSKitModels.nonsym_xxz_ladder_infinite — Function

An infinite ladder: infinite in the x direction and pbc in the y direction

Step 1: constructing a vector containing all bonds of the lattice: bonds

Step 2: summing up all two-site operators (opp): H = H + LocalOperator(opp, (bond.first,bond.second))

MPSKitModels.u1_qed_ham — Function

the lattice qed hamiltonian, using a seperate U(1) charge for even and odd matter sites, allowing one to construct manifestly guage invariant mps's