Operators

Spin operators

The spin operators sigma_x, sigma_y and sigma_z are defined such that they obey the spin commutation relations $[Sⱼ, Sₖ] = i ɛⱼₖₗ Sₗ$. Additionally, the ladder operators are defined as $S± = Sˣ ± i Sʸ$. Several combinations are defined that act on two spins.

When imposing symmetries, by convention we choose sigma_z as the diagonal operator, such that for non-trivial symmetry only the combinations that are invariant under this symmetry are implemented. As such, when defining the other single-site operators with a symmetry, an additional virtual space is required to carry the charge, which is by convention chosen as the second space in a (2,1) tensor.

MPSKitModels.sigma_x — Function

sigma_x([eltype [, symmetry]]; spin=S)

spin S operator along the x-axis.

See also σˣ

MPSKitModels.sigma_y — Function

sigma_y([eltype [, symmetry]; spin=S)

spin S operator along the y-axis.

See also σʸ

MPSKitModels.sigma_z — Function

sigma_z([eltype [, symmetry]]; spin=S)

spin S operator along the z-axis.

See also σᶻ

MPSKitModels.sigma_plus — Function

sigma_plus([eltype [, symmetry]]; spin=S)

spin S raising operator.

See also σ⁺

MPSKitModels.sigma_min — Function

sigma_min([eltype [, symmetry]]; spin=S)

spin S lowering operator.

See also σ⁻

MPSKitModels.sigma_xx — Function

sigma_xx([eltype [, symmetry]]; spin=S)

spin S xx exchange operator.

See also σˣˣ

MPSKitModels.sigma_yy — Function

sigma_yy([eltype [, symmetry]]; spin=S)

spin S yy exchange operator.

See also σʸʸ

MPSKitModels.sigma_zz — Function

sigma_zz([eltype [, symmetry]]; spin=S)

spin S zz exchange operator.

See also σᶻᶻ

MPSKitModels.sigma_plusmin — Function

sigma_plusmin([eltype [, symmetry]]; spin=S)

spin S +- exchange operator.

See also σ⁺⁻

MPSKitModels.sigma_minplus — Function

sigma_minplus([eltype [, symmetry]]; spin=S)

spin S -+ exchange operator.

See also σ⁻⁺

MPSKitModels.sigma_exchange — Function

sigma_exchange([eltype [, symmetry]]; spin=S)

spin S exchange operator.

See also σσ

For convenience, the spin 1/2 case, which reduces to the pauli matrices, have the exported unicode symbols:

MPSKitModels.σˣ — Constant

Pauli x operator

MPSKitModels.σʸ — Constant

Pauli y operator

MPSKitModels.σᶻ — Constant

Pauli z operator

MPSKitModels.σ⁺ — Constant

Pauli raising operator

MPSKitModels.σ⁻ — Constant

Pauli lowering operator

MPSKitModels.σˣˣ — Constant

Pauli xx exchange operator

MPSKitModels.σʸʸ — Constant

Pauli yy exchange operator

MPSKitModels.σᶻᶻ — Constant

Pauli zz exchange operator

MPSKitModels.σ⁺⁻ — Constant

Pauli +- exchange operator

MPSKitModels.σ⁻⁺ — Constant

Pauli -+ exchange operator

MPSKitModels.σσ — Constant

Pauli exchange operator

Bosonic operators

The bosonic creation and annihilation operators a_plus ($a^\dagger$) and a_min ($a$) are defined such that the following holds:

\[a^\dagger \left|n\right> = \sqrt(n + 1) \left|n+1\right>\]

\[a \left|n\right> = \sqrt(n) \left|n-1\right>\]

From these, a number operator $a_number$ ($N$) can be defined:

\[N = a^\dagger a\]

\[N\left|n\right> = n \left|n\right>\]

With these, the following commutators can be obtained:

\[\left[a, a^\dagger\right] = 1\]

\[\left[N,a^\dagger\right] = a^\dagger\]

\[\left[N,a\right] = -a\]

MPSKitModels.a_plus — Function

a_plus(cutoff, [elt], [symmetry])

bosonic creation operator.

MPSKitModels.a_min — Function

a_min(cutoff, [elt], [symmetry])

bosonic annihilation operator.

Missing docstring.

Missing docstring for a_number. Check Documenter's build log for details.

Fermionic operators

MPSKitModels.c_plus — Function

c_plus(elt=ComplexF64, ::Type{fℤ₂}; side=:L)

fermionic creation operator.

MPSKitModels.c_min — Function

c_min(elt=ComplexF64, ::Type{fℤ₂}; side=:L)

fermionic creation operator.

MPSKitModels.c_number — Function

c_number(elt=ComplexF64, ::Type{fℤ₂}=fℤ₂)

fermionic number operator.

MPSKitModels.e_plus — Function

e_plus(elt=Complexf64, particle_symmetry, spin_symmetry; side=:L)

creation operator for electron-like fermions.

MPSKitModels.e_min — Function

e_min(elt=Complexf64, particle_symmetry, spin_symmetry; side=:L)

annihilation operator for electron-like fermions.

MPSKitModels.e_number — Function

e_number(elt=ComplexF64, particle_symmetry=fℤ₂, spin_symmetry=ℤ₁)

number operator for electron-like fermions.