Operators

There are several different operators defined, which all follow an interface similar to the following:

operator([elt::Type{<:Number}=ComplexF64], [symmetry::Type{<:Sector}=Trivial]; kwargs...)

Here, the scalar type of the operator is defined by elt, while the symmetry can be chosen through the symmetry argument. Other parameters are supplied as keywords. The special keyword argument side can be used for operators that require an additional virtual space to satisfy the symmetry constraints, in which case it determines where this auxiliary space is located, either to the left :L (default) or to the right :R.

Spin operators

The spin operators S_x, S_y and S_z are defined such that they obey the spin commutation relations $[Sⱼ, Sₖ] = i ɛⱼₖₗ Sₗ$. Additionally, the ladder operators are defined as $S± = Sˣ ± i Sʸ$. Several combinations are defined that act on two spins.

When imposing symmetries, by convention we choose S_z as the diagonal operator for U₁, and S_x as the diagonal operator for ℤ₂.

MPSKitModels.S_x — Function

S_x([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin operator along the x-axis.

See also σˣ

MPSKitModels.S_y — Function

S_y([eltype::Type{<:Complex}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin operator along the y-axis.

See also σʸ

MPSKitModels.S_z — Function

S_z([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin operator along the z-axis.

See also σᶻ

MPSKitModels.S_plus — Function

S_plus([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin plus operator.

See also σ⁺

MPSKitModels.S_min — Function

S_min([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin minus operator.

See also σ⁻

MPSKitModels.S_xx — Function

S_xx([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin xx exchange operator.

See also σˣˣ

MPSKitModels.S_yy — Function

S_yy([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin yy exchange operator.

See also σʸʸ

MPSKitModels.S_zz — Function

S_zz([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin zz exchange operator.

See also σᶻᶻ

MPSKitModels.S_plusmin — Function

S_plusmin([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin plusmin exchange operator.

See also σ⁺⁻

MPSKitModels.S_minplus — Function

S_minplus([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin minplus exchange operator.

See also σ⁻⁺

MPSKitModels.S_exchange — Function

S_exchange([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin exchange operator.

See also σσ

For convenience, the Pauli matrices can also be recovered as $σⁱ = 2 Sⁱ$.

MPSKitModels.σˣ — Function

Pauli x operator

MPSKitModels.σʸ — Function

Pauli y operator

MPSKitModels.σᶻ — Function

Pauli z operator

MPSKitModels.σ⁺ — Function

Pauli plus operator

MPSKitModels.σ⁻ — Function

Pauli minus operator

MPSKitModels.σˣˣ — Function

Pauli xx operator

MPSKitModels.σʸʸ — Function

Pauli yy operator

MPSKitModels.σᶻᶻ — Function

Pauli zz operator

MPSKitModels.σ⁺⁻ — Function

Pauli plusmin operator

MPSKitModels.σ⁻⁺ — Function

Pauli minplus operator

MPSKitModels.σσ — Function

Pauli exchange operator

Bosonic operators

The bosonic creation and annihilation operators a_plus ($a^\dagger$) and a_min ($a$) are defined such that the following holds:

\[a^\dagger \left|n\right> = \sqrt(n + 1) \left|n+1\right>\]

\[a \left|n\right> = \sqrt(n) \left|n-1\right>\]

From these, a number operator $a_number$ ($N$) can be defined:

\[N = a^\dagger a\]

\[N\left|n\right> = n \left|n\right>\]

With these, the following commutators can be obtained:

\[\left[a, a^\dagger\right] = 1\]

\[\left[N,a^\dagger\right] = a^\dagger\]

\[\left[N,a\right] = -a\]

MPSKitModels.a_plus — Function

a_plus([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff=5)

The truncated bosonic creation operator, with a maximum of cutoff bosons per site.

MPSKitModels.a_min — Function

a_min([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff=5)

The truncated bosonic annihilation operator, with a maximum of cutoff bosons per site.

MPSKitModels.a_number — Function

a_number([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff=5)

The truncated bosonic number operator, with a maximum of cutoff bosons per site.

Fermionic operators

MPSKitModels.c_plus — Function

c_plus([elt::Type{<:Number}=ComplexF64]; side=:L)

fermionic creation operator.

MPSKitModels.c_min — Function

c_min([elt::Type{<:Number}=ComplexF64]; side=:L)

fermionic annihilation operator.

MPSKitModels.c_number — Function

c_number([elt::Type{<:Number}=ComplexF64])

fermionic number operator.

MPSKitModels.e_plus — Function

e_plus([elt::Type{<:Number}=ComplexF64], particle_symmetry, spin_symmetry; side=:L)

The creation operator for electron-like fermions.

MPSKitModels.e_min — Function

e_min([elt::Type{<:Number}=ComplexF64], particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector}; side=:L)

The annihilation operator for electron-like fermions.

MPSKitModels.e_number — Function

e_number([elt::Type{<:Number}=ComplexF64], particle_symmetry=fℤ₂, spin_symmetry=ℤ₁)

The number operator for electron-like fermions.